Isolate findFirstDebuggableLine function in proper utils file
