+++ packages/php-wasm/xdebug-bridge/src/lib/index.ts
@@ -2,4 +2,5 @@ export * from './cdp-server';
 export * from './dbgp-session';
 export * from './run-cli';
 export * from './start-bridge';
+export * from './utils';
 export * from './xdebug-cdp-bridge';
