+++ packages/playground/cli/src/run-cli.ts
@@ -49,7 +49,7 @@ import { isValidWordPressSlug } from './is-valid-wordpress-slug';
 import { resolveBlueprint } from './resolve-blueprint';
 import { BlueprintsV2Handler } from './blueprints-v2/blueprints-v2-handler';
 import { BlueprintsV1Handler } from './blueprints-v1/blueprints-v1-handler';
-import { startBridge } from '@php-wasm/xdebug-bridge';
+import { findFirstDebuggableLine, startBridge } from '@php-wasm/xdebug-bridge';
 import { basename } from 'path';
 
 export async function parseOptionsAndRunCLI() {
@@ -564,6 +564,7 @@ export async function runCLI(args: RunCLIArgs): Promise<RunCLIServer> {
 							const line = findFirstDebuggableLine(
 								await playground!.readFileAsText(filePath)
 							);
+
 							bridge.setXdebugBreakpoint(filePath, line);
 
 							await playground!.run({
@@ -743,63 +744,4 @@ async function zipSite(
 	});
 	const zip = await playground.readFileAsBuffer('/tmp/build.zip');
 	fs.writeFileSync(outfile, zip);
-}
-
-function findFirstDebuggableLine(content: string) {
-	const lines = content.split('\n');
-	let inBlockComment = false;
-	let inFunctionOrClass = false;
-	let braceDepth = 0;
-
-	for (let i = 0; i < lines.length; i++) {
-		const lineRaw = lines[i];
-		const line = lineRaw.trim();
-
-		if (line === '') continue;
-
-		if (line.startsWith('/*')) {
-			inBlockComment = true;
-			continue;
-		}
-
-		if (inBlockComment) {
-			if (line.includes('*/')) inBlockComment = false;
-			continue;
-		}
-
-		if (line.match(/^\s*(function|class)\b/)) {
-			inFunctionOrClass = true;
-		}
-
-		braceDepth += (line.match(/{/g) || []).length;
-		braceDepth -= (line.match(/}/g) || []).length;
-
-		if (inFunctionOrClass && braceDepth === 0) {
-			inFunctionOrClass = false;
-			continue;
-		}
-
-		if (inFunctionOrClass || braceDepth > 0) {
-			continue;
-		}
-
-		if (
-			line.startsWith('//') ||
-			line.startsWith('#') ||
-			line === '<?php' ||
-			line === '?>'
-		) {
-			continue;
-		}
-
-		if (
-			line.match(
-				/^\s*(var_dump|print|echo|exit|die|return|require|include|define|[$]\w+|\w+\s*\()/
-			)
-		) {
-			return i + 1;
-		}
-	}
-
-	return 0;
 }
